chore: move stopsnearyou components to a sub folder
