+++ sass/_main.scss
@@ -27,7 +27,7 @@ $body-font-weight: $font-weight-medium;
 @import '../app/component/routepage/route';
 @import '../app/component/checkbox';
 @import '../app/component/toggle';
-@import '../app/component/stops-near-you';
+@import '../app/component/stopsnearyou/stops-near-you';
 @import '../app/component/disruption';
 @import '../app/component/departure';
 @import '../app/component/stop';
